Split substrate context to separate files
