+++ packages/ui/src/hooks/useSubstrate.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { SubstrateContext } from '../providers/substrate/context'
+
+export const useSubstrate = () => ({ ...useContext(SubstrateContext) })
