+++ packages/ui/src/pages/Profile/Profile.tsx
@@ -3,8 +3,8 @@ import { KeyringAddress } from '@polkadot/ui-keyring/types'
 import { Page } from '../../components/page/Page'
 import { Navigation } from '../../components/page/Navigation'
 import { PageContent } from '../../components/page/PageContent'
-import { useSubstrate } from '../../providers/SubstrateContext'
 import { Accounts } from './Accounts'
+import { useSubstrate } from '../../hooks/useSubstrate'
 
 export function Profile() {
   const { keyring, keyringState } = useSubstrate()
